Resolved issue, forgot that switched from double to single precision, thus did not reach same residual. Make sure to consider that in the future. Else, parallel code works fine

+++ minSurf.cpp
@@ -33,7 +33,7 @@
 
 int main() {
 
-    typedef float dType;
+    typedef double dType;
     typedef Eigen::Matrix<dType,-1, 1> mType; // -1 -> Dynamic allocation (for arrays > 16, what 
                                                // we usually have, there is no performance benefit 
                                                // for fixed size allocation
